Refactoring. Especially in the competition.utils.interactions.groups package
