+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/competition/utils/interactions/groups/StandardCarfax.java
@@ -3,7 +3,7 @@ package org.firstinspires.ftc.teamcode.competition.utils.interactions.groups;
 import org.firstinspires.ftc.teamcode.competition.utils.interactions.InteractionSurface;
 import org.firstinspires.ftc.teamcode.competition.utils.interactions.items.StandardMotor;
 
-public class StandardCarfax extends InteractionSurface {
+public class StandardCarfax extends DrivetrainTemplate {
 
     private final StandardMotor RIGHT, LEFT;
